refactor(serialization): move word implementation module
