+++ zk_ee/src/oracle/word_serialization/dyn_word_iterator.rs
