+++ zk_ee/src/oracle/word_serialization/tests.rs
